+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.0.3
+
+* Add handling for some kind of variadic arguments handling.
+
+  A command or function will be passed `Nothing` as it's
+  last arguments if the argument type is wrapped in `Maybe`
+  and the invocation on the side of neovim did not pass those
+  arguments.
+
 # 0.0.2
 
 * Add handling for special command options